Merge convergence clubs

Merges a list of clubs created with the function findClubs by either Phillips and Sul method or von Lyncker and Thoennessen procedure.

mergeClubs(clubs, time_trim, mergeMethod = c("PS", "vLT"),
  mergeDivergent = FALSE, threshold = -1.65)

Arguments

clubs
an object of class convergence.clubs (created by findClubs function)

time_trim
a numeric value between 0 and 1, representing the portion of time periods to trim when running log t regression model; if omitted, the same value used for clubs is used.

mergeMethod
character string indicating the merging method to use. Methods available are "PS" for Phillips and Sul (2009) and "vLT" for von Lyncker and Thoennessen (2016).

mergeDivergent
logical, if TRUE, indicates that merging of divergent regions should be tried.

threshold
a numeric value indicating the threshold to be used with the t-test.

Value

Ad object of class convergence.clubs, containing a list of Convergence Clubs, for each club a list is return with the following objects: id, a vector containing the row indices of the regions in the club; model, a list containing information about the model used to run the t-test on the regions in the club; regions, a vector containing the names of the regions of the club (optional, only included if parameter regions is given)

Details

Phillips and Sul (2009) suggest a "club merging algorithm" to avoid over determination due to the selection of the parameter \(c^*\). This algorithm suggests to merge for adjacent groups. In particular, it works as follows:

Take the first two groups detected in the basic clustering mechanism and run the log-t test. If the t-statistic is larger than -1.65, these groups together form a new convergence club;

Repeat the test adding the next group and continue until the basic condition (t-statistic > -1.65) holds;

If convergence hypothesis is rejected, conclude that all previous groups converge, except the last one. Hence, start again the test merging algorithm beginning from the group for which the hypothesis of convergence did not hold. On the other hand, von Lyncker and Thoennessen (2016), propose a modified version of the club merging algorithm that works as follows:

Take all the groups detected in the basic clustering mechanism (P) and run the t-test for adjacent groups, obtaining a (M × 1) vector of convergence test statistics t (where \(M = P - 1\) and \(m = 1, \dots, M\));

Merge for adjacent groups starting from the first, under the conditions \(t(m) > -1.65\) and \(t(m) > t(m+1)\). In particular, if both conditions hold, the two clubs determining \(t(m)\) are merged and the algorithm starts again from step 1, otherwise it continues for all following pairs;

For the last element of vector M (the value of the last two clubs) the only condition required for merging is \(t(m=M) > -1.65\).

References

Phillips, P. C.; Sul, D., 2007. Transition modeling and econometric convergence tests. Econometrica 75 (6), 1771-1855.

Phillips, P. C.; Sul, D., 2009. Economic transition and growth. Journal of Applied Econometrics 24 (7), 1153-1185.

von Lyncker, K.; Thoennessen, R., 2016. Regional club convergence in the EU: evidence from a panel data analysis. Empirical Economics.

See also

Examples

data("countryGDP")

# Cluster Countries using GDP from year 1970 to year 2003
clubs <- findClubs(countryGDP, dataCols=2:35, regions = 1, refCol=35,
                   time_trim = 1/3, cstar = 0, HACmethod = "FQSB")
summary(clubs)

#> Number of convergence clubs: 7
#> Number of divergent units: 0
#> 
#>          # of regions 
#> ------- ---------------
#> club1  | 50           
#> club2  | 30           
#> club3  | 21           
#> club4  | 24           
#> club5  | 14           
#> club6  | 11           
#> club7  | 2            

# Merge clusters
mclubs <- mergeClubs(clubs, mergeMethod='PS', mergeDivergent=FALSE)
summary(mclubs)

#> Number of convergence clubs: 6
#> Number of divergent units: 0
#> 
#>          # of regions  | merged clubs 
#> ------- --------------- --------------
#> club1  | 50            | clubs: 1     
#> club2  | 30            | clubs: 2     
#> club3  | 21            | clubs: 3     
#> club4  | 38            | clubs: 4, 5  
#> club5  | 11            | clubs: 6     
#> club6  | 2             | clubs: 7     

mclubs <- mergeClubs(clubs, mergeMethod='vLT', mergeDivergent=FALSE)
summary(mclubs)

#> Number of convergence clubs: 6
#> Number of divergent units: 0
#> 
#>          # of regions  | merged clubs 
#> ------- --------------- --------------
#> club1  | 50            | clubs: 1     
#> club2  | 30            | clubs: 2     
#> club3  | 21            | clubs: 3     
#> club4  | 38            | clubs: 4, 5  
#> club5  | 11            | clubs: 6     
#> club6  | 2             | clubs: 7