Merge divergent units

Merges divergent units according the algorithm proposed by von Lyncker and Thoennessen (2016)

mergeDivergent(clubs, time_trim, threshold = -1.65)

Arguments

clubs
an object of class convergence.clubs (created by findClub or mergeClubs function)

time_trim
a numeric value between 0 and 1, representing the portion of time periods to trim when running log t regression model; if omitted, the same value used for clubs is used.

threshold
a numeric value indicating the threshold to be used with the t-test.

Value

A list of Convergence Clubs, for each club a list is return with the following objects: id, a vector containing the row indices of the regions in the club; model, a list containing information about the model used to run the t-test on the regions in the club; regions, a vector containing the names of the regions of the club (optional, only included if it is present in the clubs object given in input).

Details

von Lyncker and Thoennessen (2016) claim that units identified as divergent by the basic clustering procedure by Phillips and Sul might not necessarily still diverge in the case of new convergence clubs detected with the club merging algorithm. To test if divergent regions may be included in one of the new convergence clubs, they propose the following algorithm:

Run a log t-test for all diverging regions, and if \(t_k > -1.65\) all these regions form a convergence club (This step is implicitly included in Phillips and Sul basic algorithm);

Run a log t-test for each diverging regions and each club, creating a matrix of t-values with dimensions \(d \times p\), where each row d represents a divergent region and each column p a convergence club;

Take the highest \(t > e^*\) and add the respective region to the respective club and restart from the step 1. the authors suggest to use \(e^* = t = -1.65 \);

The algorithm stops when no t-value > e* is found in step 3, and as a consequence all remaining regions are considered divergent.

References

Phillips, P. C.; Sul, D., 2007. Transition modeling and econometric convergence tests. Econometrica 75 (6), 1771-1855.

Phillips, P. C.; Sul, D., 2009. Economic transition and growth. Journal of Applied Econometrics 24 (7), 1153-1185.

von Lyncker, K.; Thoennessen, R., 2016. Regional club convergence in the EU: evidence from a panel data analysis. Empirical Economics, doi:10.1007/s00181-016-1096-2, 1-29.

See also

Examples

data("countryGDP")

#Cluster Countries using GDP from year 1970 to year 2003
clubs <- findClubs(countryGDP, dataCols=2:35, regions = 1, refCol=35,
                   time_trim = 1/3, cstar = 0, HACmethod = "AQSB")
summary(clubs)

#> Number of convergence clubs: 5
#> Number of divergent units: 3
#> 
#>              # of regions 
#> ----------- ---------------
#> club1      | 84           
#> club2      | 18           
#> club3      | 26           
#> club4      | 14           
#> club5      | 7            
#> divergent  | 3            

# Merge clusters and divergent regions
mclubs <- mergeClubs(clubs, mergeDivergent=TRUE)
summary(mclubs)

#> Number of convergence clubs: 4
#> Number of divergent units: 1
#> 
#>              # of regions  | merged clubs 
#> ----------- --------------- --------------
#> club1      | 84            | clubs: 1     
#> club2      | 18            | clubs: 2     
#> club3      | 40            | clubs: 3, 4  
#> club4      | 9             | clubs: 5     
#> divergent  | 1             | clubs: